Moved from manager package since it actually does not manage something.
